+++ app/src/main/java/com/mobile/app_iara/ui/gestao/GestaoFragment.kt
@@ -1,4 +1,4 @@
-package com.mobile.app_iara.ui.notifications
+package com.mobile.app_iara.ui.gestao
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.mobile.app_iara.databinding.FragmentNotificationsBinding
 
-class NotificationsFragment : Fragment() {
+class GestaoFragment : Fragment() {
 
     private var _binding: FragmentNotificationsBinding? = null
 
@@ -23,7 +23,7 @@ class NotificationsFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         val notificationsViewModel =
-            ViewModelProvider(this).get(NotificationsViewModel::class.java)
+            ViewModelProvider(this).get(GestaoViewModel::class.java)
 
         _binding = FragmentNotificationsBinding.inflate(inflater, container, false)
         val root: View = binding.root
