+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -19,9 +19,9 @@
 
     <fragment
         android:id="@+id/navigation_gestao"
-        android:name="com.mobile.app_iara.ui.notifications.NotificationsFragment"
+        android:name="com.mobile.app_iara.ui.gestao.GestaoFragment"
         android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        tools:layout="@layout/fragment_gestao" />
     <fragment
         android:id="@+id/navigation_perfil"
         android:name="com.mobile.app_iara.ui.profile.ProfileFragment"
